% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixes.r
\name{fix_eurocontrol_region}
\alias{fix_eurocontrol_region}
\title{Fix region value for various countries according to EUROCONTROL classification.}
\usage{
fix_eurocontrol_region(flows)
}
\arguments{
\item{flows}{a data frame with origin/destination flows}
}
\value{
a data frame with corrected region
}
\description{
Fix region value for various countries according to EUROCONTROL classification.
}
\examples{
\dontrun{
flows \%>\% fix_eurocontrol_region()
}
}
