% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixes.r
\name{fix_iso_countrycode}
\alias{fix_iso_countrycode}
\title{Fix ICO2 code}
\usage{
fix_iso_countrycode(flows)
}
\arguments{
\item{flows}{a data frame with origin/destination flows}
}
\value{
a data frame with corrected ISO2 codes
}
\description{
Some ISO2 codes in the EUROCONTROL data are wrong.
For example Serbia is "RS" since 2006, not anymore "CS" linke in our data.
}
\examples{
\dontrun{
flows \%>\% fix_iso_countrycode()
}

}
