% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixes.r
\name{fix_region_AN}
\alias{fix_region_AN}
\title{Fix "AN" region value to "Caribbean"}
\usage{
fix_region_AN(flows)
}
\arguments{
\item{flows}{a data frame with origin/destination flows}
}
\value{
a data frame with corrected region
}
\description{
Fix "AN" region value to "Caribbean"
}
\examples{
\dontrun{
flows \%>\% fix_region_AN()
}
}
