% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixes.r
\name{fix_region_IO}
\alias{fix_region_IO}
\title{Fix "IO" region value to "Eastern Africa"}
\usage{
fix_region_IO(flows)
}
\arguments{
\item{flows}{a data frame with origin/destination flows}
}
\value{
a data frame with corrected region
}
\description{
Fix "IO" region value to "Eastern Africa"
}
\examples{
\dontrun{
flows \%>\% fix_region_IO()
}
}
