% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixes.r
\name{fix_region_TW}
\alias{fix_region_TW}
\title{Fix "TW" region value to "Eastern Asia"}
\usage{
fix_region_TW(flows)
}
\arguments{
\item{flows}{a data frame with origin/destination flows}
}
\value{
a data frame with corrected region
}
\description{
Fix "TW" region value to "Eastern Asia"
}
\examples{
\dontrun{
flows \%>\% fix_region_TW()
}
}
