% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifrflows-package.r
\docType{package}
\name{ifrflows}
\alias{ifrflows}
\title{Tools for Creating flows of flights data}
\description{
There are functions to export data from PRIME and all kind of tricks to clean the mess.
}
\author{
Enrico Spinielli (@espinielli)
}
