% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.r
\name{retrieve_airport_flows}
\alias{retrieve_airport_flows}
\title{retrieve the yearly number of flights between airports for a period of years.}
\usage{
retrieve_airport_flows(wef, til)
}
\arguments{
\item{wef}{the beginning year of the period (included).}

\item{til}{the end year of the period (not included).}
}
\value{
a dataframe with year, # of flights, departing country code and arrival country code.
}
\description{
retrieve the yearly number of flights between airports for a period of years.
}
\examples{
\dontrun{
retrieve_airport_flows('2001', '2002')
# A tibble: 8,093,679 x 6
YEAR  NB_OF_FLIGHT ADEP  DEP_ISO_COUNTRY_CODE ADES  DES_ISO_COUNTRY_CODE
<chr>        <dbl> <chr> <chr>                <chr> <chr>
  1 2001           14. EGKB  GB                   LFMD  FR
2 2001           12. LFMD  FR                   LGKR  GR
3 2001            9. EGKB  GB                   LFMD  FR
4 2001            9. LEMD  ES                   LEBL  ES
5 2001            7. EDDR  DE                   ELLX  LU
6 2001            6. LEBL  ES                   LEMD  ES
7 2001            6. LGAT  GR                   LGTS  GR
8 2001            5. EHBK  NL                   EHBK  NL
9 2001            5. GMMX  MA                   LFPB  FR
10 2001            5. EGGW  GB                   EGSS  GB
# ... with 8,093,669 more rows
}

}
