% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{retrieve_airports}
\alias{retrieve_airports}
\title{Retrieve airport details}
\usage{
retrieve_airports(apts, wef, til)
}
\arguments{
\item{apts}{a The ICAO airport IDs to retrieve information about (TODO)}

\item{wef}{the beginning date of the period (included).}

\item{til}{the end date of the period (not included).}
}
\value{
A dataframe with
\itemize{
  \item icao: ICAO ID
  \item longitude: Airport Reference Point (APR) longitude
  \item latitude: Airport Reference Point (APR) latitude
  \item name
}
}
\description{
Retrieve airport details
}
\examples{
\dontrun{
retrieve_airports("EBBR", "2002-11-24", "2003-01-30")
}

}
