% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_daily_airport_flows.R
\name{retrieve_daily_airport_flows}
\alias{retrieve_daily_airport_flows}
\title{Extract daily flight country flows}
\usage{
retrieve_daily_airport_flows(wef, til)
}
\arguments{
\item{wef}{Start date of the interval of interest (format "YYYY-MM-DD")}

\item{til}{End date of the interval of interest (NOTE: non-inclusive; format "YYYY-MM-DD")}
}
\value{
A tibble
}
\description{
Extract country pair flows of flights for the specified date interval.
}
\details{
You need access to PRISME's PRU_DEV schema to retrieve the dataset.
         ISO country codes need scrutiny because they are not currently well
         maintained in PRISME, i.e. "CS" for Serbia.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
